+++ storage/src/backend/mod.rs
@@ -3,10 +3,14 @@
 
 pub mod btree_map;
 pub mod in_memory_backend;
+pub mod mark_move_gced;
+pub mod mark_sweep_gced;
 pub mod rocksdb_backend;
 pub mod sled_backend;
 
 pub use btree_map::*;
 pub use in_memory_backend::*;
+pub use mark_sweep_gced::*;
+pub use mark_sweep_gced::*;
 pub use rocksdb_backend::*;
 pub use sled_backend::*;